Organize frontend folder, and add user type
