+++ frontend/src/App.tsx
@@ -3,11 +3,11 @@ import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
 import { BrowserRouter,
   Routes, Route } from "react-router-dom";
-import {LoginPage} from "./components/routes/LoginPage";
-import { Home } from "./components/routes/HomePage";
-import { RegisterPage } from "./components/routes/RegisterPage";
-import { AuthProvider } from "./components/authentication/AuthProvider";
-import { PastTradesPage } from "./components/routes/PastTradesPage";
+import {LoginPage} from "./routes/LoginPage";
+import { Home } from "./routes/HomePage";
+import { RegisterPage } from "./routes/RegisterPage";
+import { AuthProvider } from "./contexts/authentication/AuthProvider";
+import { PastTradesPage } from "./routes/PastTradesPage";
 import { Nav } from "./components/nav/Nav";
 
 function App() {
