+++ frontend/src/components/user/LoginForm.tsx
@@ -2,7 +2,7 @@ import "bootstrap/dist/css/bootstrap.css";
 import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import { useState, useContext } from "react";
-import { AuthContext } from "../authentication/AuthContext";
+import { AuthContext } from "../../contexts/authentication/AuthContext";
 
 export const LoginForm = (props: any) => {
   const [email, setEmail] = useState("");
