+++ frontend/src/contexts/authentication/RequireAuth.tsx
